requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document Processing (Phase 3-4)
pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10

# AI/LLM Integration (Phase 4)
openai==1.6.1

# Testing
pytest==7.4.3
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2